added bullet damage

+++ Global/PlayerStats.cs
@@ -17,7 +17,7 @@ public partial class PlayerStats : Node
 	public Vector2[] hardpoint_locations;
 	public string[] hardpoint_weight_classes;
 
-	public double ship_width;
+	public int ship_width;
 	public Vector2 ship_start_point;
 	public string current_ship_UID;
 	public override void _Ready()
